refactor: split out styles and theme
