+++ src/theme/index.ts
@@ -1,0 +1,3 @@
+export * from './defaultTheme';
+export * from './styles';
+export * from './theme';
